Scrape sparkline URL instead of constructing it
For custom domains, if the admin area is visited and the custom domain
has not been visited, no cookie for the custom domain will be present
and any requests to the custom domain will be as if the user is not
logged in.

The sparkline image src is already available, so just scape it and
modify it. This introduces a vulnerability to upstream layout changes,
but it is the most reliable method.